requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlite3
asyncio
datetime
typing
hashlib
secrets
logging
